pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexa"
version = "0.1.0"
description = "Nexa workflow engine"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "fastapi>=0.111",
  "httpx>=0.27"
]

[project.scripts]
nexa = "src.cli.nexa_cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
[project.optional-dependencies]
aws = ["boto3>=1.34"]
testing = ["testcontainers[postgres]>=4.0", "pytest-asyncio>=0.23"]
